refactor : BookItemType 선언 위치 변경
