+++ src/types/index.ts
@@ -1,3 +1,5 @@
+export * from "./book-item.type"
+
 /** [공용 타입 관리 폴더]
 ---------------------------------------------------------------------------
 ## description
